db_assistant_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.33.0
trino==0.326.0
python-dotenv==1.0.1
pydantic==2.7.1
